+++ src/xyz/oihana/schema/constants/traits/AuthTrait.php
@@ -4,6 +4,7 @@ namespace xyz\oihana\schema\constants\traits;
 
 use xyz\oihana\schema\constants\traits\auth\ApplicationTrait;
 use xyz\oihana\schema\constants\traits\auth\InvitationTrait;
+use xyz\oihana\schema\constants\traits\auth\KeyfileTrait;
 use xyz\oihana\schema\constants\traits\auth\PermissionTrait;
 use xyz\oihana\schema\constants\traits\auth\PolicyTrait;
 use xyz\oihana\schema\constants\traits\auth\RoleTrait;
@@ -15,6 +16,7 @@ use xyz\oihana\schema\constants\traits\auth\WebAPITrait;
 trait AuthTrait
 {
     use ApplicationTrait ,
+        KeyfileTrait ,
         InvitationTrait ,
         PermissionTrait ,
         PolicyTrait ,